redcoil fix, no more shredding of blocks with the "Any" prefix, GUIs
